requirements.txt
marshmallow-enum==1.5.1
marshmallow==2.16.0
djangorestframework==3.9.4
django==1.11.17
six==1.15.0
errand-boy==0.3.9
